cert-inspection: move functions to separate utils pkg
